stakeholder preview card
- clean up some code to improve legibility
- remove stakeholder preview from ResultsList and make it it's own reusable component
